much better coincidence finding (works for more than 3 stations, works over seconds

+++ czelta/event.h
@@ -37,10 +37,10 @@ public:
     Event(time_t timestamp,double last_second, int16_t TDC0, int16_t TDC1, int16_t TDC2, int16_t ADC0, int16_t ADC1, int16_t ADC2, int16_t t0, int16_t t1, int16_t t2, int8_t tCrateRaw, bool calibration, uint8_t station);
     Event(const Event& orig);
     bool operator==(const Event&) const;
-    inline int timestamp() const{return _timestamp;};
+    inline uint32_t timestamp() const{return _timestamp;};
     inline double last_second() const{return _last_second*1e-1;}; //time since last second, in nanoseconds
     inline double time_since_second() const{return _last_second*1e-10;};//time since last second, in seconds
-    inline int64_t tenthOfNSTimestamp() const{return int64_t(timestamp())*10000000000L + _last_second;};
+    inline __int128 tenthOfNSTimestamp() const{return __int128(timestamp())*10000000000L + _last_second;};
     inline short TDC0() const{return _TDC0;};
     inline short TDC1() const{return _TDC1;};
     inline short TDC2() const{return _TDC2;};
